Intranet Stock Order System
Clara has access to this, and uses it to make requests for new stock, depending on what the territory leads have said. There is also a request system that the In-home Teams outside of Auckland use. The Auckland ones prefer to just talk to Clara directly. Clara has offered me read only access, so I will see how that goes.